Created the ability to click on a movie and display more details
Added an Overview component. This component is a Modal that displays when a user clicks on a movie poster. The Modal displays information about the movie that was selected.

Material-ui was added as a dependency to be able to use the Modal it provides.

I added more comments on areas that require them.