Fixed a potential page state issue.
While reading more about hooks, I learned that if you make multiple changes to a state within a setState you might run into an issue where the state isn't updated correctly. This is due to the fact that the you're making a change to a state when it was rendered. To fix this we want to make a function within the setState that passes in the previous state passed in. This resolves any issues.