Mitigate the Search bar API requests
Previously the Search bar would make an API request anytime a keystroke was made in the search bar. We want to limit the amount of requests if we can so I put a debouncer in the Search component. This will set a timeout for each keystroke.

I also separated the getImageData fetch into its own useEffect. This way we are not requesting image data each time we update the state of page, filter, or query.